backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
langgraph==0.2.60
langchain==0.3.13
langchain-core==0.3.28
litellm==1.56.4
tavily-python==0.5.0
python-dotenv==1.0.1
httpx>=0.23.0,<0.28.0
